Refactor: extract remove filter functions
